+++ astrality/tests/module/test_configuration.py
@@ -20,7 +20,8 @@ def test_modules_config_explicitly_enabled_modules(
         config=global_modules_config_dict,
         config_directory=test_config_directory,
     )
-    global_modules_config.compile_config_files({})
+    for source in global_modules_config.external_module_sources:
+        source.modules({})
 
     assert 'burma::burma' in global_modules_config.enabled_modules
     assert 'india' in global_modules_config.enabled_modules
@@ -41,7 +42,8 @@ def test_modules_config_implicitly_enabled_modules(
         config=global_modules_config_dict,
         config_directory=test_config_directory,
     )
-    global_modules_config.compile_config_files({})
+    for source in global_modules_config.external_module_sources:
+        source.modules({})
 
     assert 'burma::burma' in global_modules_config.enabled_modules
     assert 'india' in global_modules_config.enabled_modules
@@ -63,7 +65,8 @@ def test_modules_config_several_implicitly_enabled_modules(
         config=global_modules_config_dict,
         config_directory=test_config_directory,
     )
-    global_modules_config.compile_config_files({})
+    for source in global_modules_config.external_module_sources:
+        source.modules({})
 
     assert 'two_modules::bhutan' in global_modules_config.enabled_modules
     assert 'two_modules::bangladesh' in global_modules_config.enabled_modules